chore - #18 response 폴더 네이밍 request로 변경
