+++ src/main/java/sopt/org/HMH/global/auth/social/kakao/fegin/KakaoApiClient.java
@@ -4,7 +4,7 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
-import sopt.org.HMH.global.auth.social.kakao.response.KakaoUserRequest;
+import sopt.org.HMH.global.auth.social.kakao.request.KakaoUserRequest;
 
 @FeignClient(name = "kakaoApiClient", url = "${oauth2.kakao.base-url}")
 public interface KakaoApiClient {
