+++ src/main/java/sopt/org/HMH/global/auth/social/kakao/request/KakaoUserProfile.java
@@ -1,11 +1,11 @@
-package sopt.org.HMH.global.auth.social.kakao.response;
+package sopt.org.HMH.global.auth.social.kakao.request;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoUserRequest(
-        Long id,
-        KakaoAccount kakaoAccount
+public record KakaoUserProfile(
+        String nickname,
+        String profileImageUrl
 ) {
 }
